Models/requirements.txt
# Core Python
numpy>=1.23
scipy>=1.10

# PyTorch (install CUDA-specific version separately if needed)
torch>=2.0
torchvision>=0.15
torchaudio>=2.0

# nnU-Net ecosystem
nnunetv2>=2.2
batchgenerators>=0.25

# Image processing
Pillow>=9.5
scikit-image>=0.21
opencv-python>=4.8

# Data handling & utilities
tqdm>=4.66
pyyaml>=6.0
einops>=0.7

# Machine learning utilities
scikit-learn>=1.3

# Medical imaging formats
SimpleITK>=2.3
nibabel>=5.1
pydicom>=2.4

# Generative models / advanced NN tools
monai>=1.3
torchmetrics>=1.2

# Physics-informed / scientific computing
sympy>=1.12
jax>=0.4; platform_system != "Windows"
jaxlib>=0.4; platform_system != "Windows"

# Logging & experiment tracking
tensorboard>=2.14
matplotlib>=3.7
seaborn>=0.13